Add files via upload
This lab is an introduction to creation of basic constraints with Xilinx Vivado. Basic constraints include specification of pin placements, pin signaling standards, and synchronous design performance requirements. No new logic design concepts are presented in this lab. The goal of this lab is for you to become more familiar with the tools. Please read carefully, pay attention, and take your time. This lab is not a race to see who gets done first.
In order to receive credit for this lab, you must complete the tasks and submit your project to the instructor. There is no in-class demonstration requirement.